requirements.txt
asgiref==3.5.0
Django==3.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.13.1
engineering-notation==0.6.0
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.9.2
pytz==2022.1
sqlparse==0.4.2
tk-tools==0.15.0
